feature offline mode. (#12)
*  - hover doesn't update on every run
  - fixes: https://github.com/go-flutter-desktop/go-flutter/issues/169
 - provide a `hover upgrade `command (like flutter upgrade)
 - every 2 days a check is made to see if the current project tags uses the last
   github release. if a new release is available print a message about
   `hover upgrade`
 - adds more debug message, 'bundling flutter', 'compiling
   go-flutter', 'running project',...
 - only print the 'channel beta' message if the user is on master,
  if the user was on stable the message was printed for no reason.

The behaviour of `hover run` and `hover build` is the same as before
when `--branch` is provided.
But there is a little twist, if the current version is a valid release
(available on github), `hover run` will not overwrite the 'go-flutter'
version. Example:

```
$ hover run --branch '@v0.20.0'
$ hover run # no upgrade trigged running on v0.20.0, at some point you
            # will get a `hover upgrade` notice

$ hover upgrade # user has to run to upgrade away from v0.20.0

$

$ hover run --branch '@master'
$ hover run # auto upgrade trigged! Not a real release.
            # If the user hasn't access to internet, fail silently and
            # keep using '@master'
```

To check every two days, a hidden file `./desktop/.last_goflutter_check`
is created on a project bases, the file contains a timestamps indicating
the last hover update check.

About the hover repo:
A empty go.mod file has been added to hover/assets/app/ this prevent adding
`go-flutter` as a go mod dependency. (because `go mod tidy` was peeking
listing hover/asset/options.go as a hover code file)

* chore: add last_goflutter_check into gitignore

* feat:run omit building flutter project or embedder

When using hover only for building plugin the `flutter build bundle`
slows down the process. Add a flag to omit `flutter build bundle`.

When making changes to the flutter package only, omit `go build
'go-flutter'` with the flag omit-embedder.

* doc: expose the branch hover flag

* fix runcmd and add observatory-port

* ecosystem consideration: prefixing files with 'go-' (#13)

* chore: add last_goflutter_check into gitignore

* ecosystem consideration: prefixing files with 'go-'

* add --path flag to all commands

* chore: fix indent

* Update init.go

* address the PR comments

* address PR comment

* move enginecache func to new versioncheck package

* updates the hover's gitignore file when migrating

adds the '.last_goflutter_check' entry to the go/.gitignore file if
needed

* internal/versioncheck: (chore) Renamed utility function

+++ cmd/build.go
@@ -9,29 +9,34 @@ import (
 	"runtime"
 	"strings"
 
+	version "github.com/hashicorp/go-version"
 	"github.com/otiai10/copy"
 	"github.com/spf13/cobra"
 
 	"github.com/go-flutter-desktop/hover/internal/enginecache"
+	"github.com/go-flutter-desktop/hover/internal/versioncheck"
 )
 
 var dotSlash = string([]byte{'.', filepath.Separator})
 
 var (
-	buildTarget    string
-	buildManifest  string
-	buildBranch    string
-	buildDebug     bool
-	buildCachePath string
+	buildTarget            string
+	buildManifest          string
+	buildBranch            string
+	buildDebug             bool
+	buildCachePath         string
+	buildOmitEmbedder      bool
+	buildOmitFlutterBundle bool
 )
+
+const buildPath = "go"
 
 func init() {
 	buildCmd.Flags().StringVarP(&buildTarget, "target", "t", "lib/main_desktop.dart", "The main entry-point file of the application.")
 	buildCmd.Flags().StringVarP(&buildManifest, "manifest", "m", "pubspec.yaml", "Flutter manifest file of the application.")
-	buildCmd.Flags().StringVarP(&buildBranch, "branch", "b", "", "The go-flutter-desktop/go-flutter branch to use when building the embedder. (@master for example)")
+	buildCmd.Flags().StringVarP(&buildBranch, "branch", "b", "", "The 'go-flutter' version to use. (@master or @v0.20.0 for example)")
 	buildCmd.Flags().BoolVar(&buildDebug, "debug", false, "Build a debug version of the app.")
 	buildCmd.Flags().StringVarP(&buildCachePath, "cache-path", "", "", "The path that hover uses to cache dependencies such as the Flutter engine .so/.dll (defaults to the standard user cache directory)")
-	buildCmd.Flags().MarkHidden("branch")
 	rootCmd.AddCommand(buildCmd)
 }
 
@@ -50,7 +55,7 @@ var buildCmd = &cobra.Command{
 }
 
 func build(projectName string, targetOS string, vmArguments []string) {
-	outputDirectoryPath, err := filepath.Abs(filepath.Join("desktop", "build", "outputs", targetOS))
+	outputDirectoryPath, err := filepath.Abs(filepath.Join(buildPath, "build", "outputs", targetOS))
 	if err != nil {
 		fmt.Printf("hover: Failed to resolve absolute path for output directory: %v\n", err)
 		os.Exit(1)
@@ -77,10 +82,13 @@ func build(projectName string, targetOS string, vmArguments []string) {
 		engineCachePath = enginecache.ValidateOrUpdateEngine(targetOS)
 	}
 
-	err = os.RemoveAll(outputDirectoryPath)
-	if err != nil {
-		fmt.Printf("hover: failed to clean output directory %s: %v\n", outputDirectoryPath, err)
-		os.Exit(1)
+	if !buildOmitFlutterBundle && !buildOmitEmbedder {
+		err = os.RemoveAll(outputDirectoryPath)
+		fmt.Printf("hover: Cleaning the build directory\n")
+		if err != nil {
+			fmt.Printf("hover: failed to clean output directory %s: %v\n", outputDirectoryPath, err)
+			os.Exit(1)
+		}
 	}
 
 	err = os.MkdirAll(outputDirectoryPath, 0775)
@@ -92,18 +100,19 @@ func build(projectName string, targetOS string, vmArguments []string) {
 	cmdCheckFlutter := exec.Command(flutterBin, "--version")
 	cmdCheckFlutterOut, err := cmdCheckFlutter.Output()
 	if err != nil {
-		fmt.Printf("hover: failed to check your flutter version: %v\n", err)
+		fmt.Printf("hover: failed to check your flutter channel: %v\n", err)
 	} else {
 		re := regexp.MustCompile("•\\schannel\\s(\\w*)\\s•")
 
 		match := re.FindStringSubmatch(string(cmdCheckFlutterOut))
 		if len(match) >= 2 {
-			if match[1] != "beta" {
+			ignoreWarning := os.Getenv("HOVER_IGNORE_CHANNEL_WARNING")
+			if match[1] != "beta" && ignoreWarning != "true" {
 				fmt.Println("hover: ⚠ The go-flutter project tries to stay compatible with the beta channel of Flutter.")
 				fmt.Println("hover: ⚠     It's advised to use the beta channel. ($ flutter channel beta)")
 			}
 		} else {
-			fmt.Printf("hover: failed to check your flutter version: Unrecognized output format")
+			fmt.Printf("hover: failed to check your flutter channel: Unrecognized output format")
 		}
 	}
 
@@ -120,10 +129,14 @@ func build(projectName string, targetOS string, vmArguments []string) {
 	)
 	cmdFlutterBuild.Stderr = os.Stderr
 	cmdFlutterBuild.Stdout = os.Stdout
-	err = cmdFlutterBuild.Run()
-	if err != nil {
-		fmt.Printf("hover: Flutter build failed: %v\n", err)
-		os.Exit(1)
+
+	if !buildOmitFlutterBundle {
+		fmt.Printf("hover: Bundling flutter app\n")
+		err = cmdFlutterBuild.Run()
+		if err != nil {
+			fmt.Printf("hover: Flutter build failed: %v\n", err)
+			os.Exit(1)
+		}
 	}
 
 	var engineFile string
@@ -163,12 +176,17 @@ func build(projectName string, targetOS string, vmArguments []string) {
 	}
 
 	err = copy.Copy(
-		filepath.Join("desktop", "assets"),
+		filepath.Join(buildPath, "assets"),
 		filepath.Join(outputDirectoryPath, "assets"),
 	)
 	if err != nil {
-		fmt.Printf("hover: Failed to copy desktop/assets: %v\n", err)
+		fmt.Printf("hover: Failed to copy %s/assets: %v\n", buildPath, err)
 		os.Exit(1)
+	}
+
+	if buildOmitEmbedder {
+		// Omit the 'go-flutter' build
+		return
 	}
 
 	var cgoLdflags string
@@ -190,33 +208,43 @@ func build(projectName string, targetOS string, vmArguments []string) {
 		os.Exit(1)
 	}
 
-	cmdGoGetU := exec.Command(goBin, "get", "-u", "github.com/go-flutter-desktop/go-flutter"+buildBranch)
-	cmdGoGetU.Dir = filepath.Join(wd, "desktop")
-	cmdGoGetU.Env = append(os.Environ(),
-		"GO111MODULE=on",
-		"CGO_LDFLAGS="+cgoLdflags,
-	)
-	cmdGoGetU.Stderr = os.Stderr
-	cmdGoGetU.Stdout = os.Stdout
+	if buildBranch == "" {
 
-	err = cmdGoGetU.Run()
-	if err != nil {
-		fmt.Printf("hover: Updating go-flutter to latest version failed: %v\n", err)
-		os.Exit(1)
-	}
+		currentTag, err := versioncheck.CurrentGoFlutterTag(filepath.Join(wd, buildPath))
+		if err != nil {
+			fmt.Printf("hover: %v\n", err)
+			os.Exit(1)
+		}
 
-	cmdGoModDownload := exec.Command(goBin, "mod", "download")
-	cmdGoModDownload.Dir = filepath.Join(wd, "desktop")
-	cmdGoModDownload.Env = append(os.Environ(),
-		"GO111MODULE=on",
-	)
-	cmdGoModDownload.Stderr = os.Stderr
-	cmdGoModDownload.Stdout = os.Stdout
+		semver, err := version.NewSemver(currentTag)
+		if err != nil {
+			fmt.Printf("hover: faild to parse 'go-flutter' semver: %v\n", err)
+			os.Exit(1)
+		}
 
-	err = cmdGoModDownload.Run()
-	if err != nil {
-		fmt.Printf("hover: Go mod download failed: %v\n", err)
-		os.Exit(1)
+		if semver.Prerelease() != "" {
+			fmt.Printf("hover: Upgrade 'go-flutter' to the latest release\n")
+			// no buildBranch provided and currentTag isn't a release,
+			// force update. (same behaviour as previous version of hover).
+			err = upgradeGoFlutter(targetOS, engineCachePath)
+			if err != nil {
+				// the upgrade can fail silently
+				fmt.Printf("hover: Upgrade ignored, current 'go-flutter' version: %s\n", currentTag)
+			}
+		} else {
+			// when the buildBranch is empty and the currentTag is a release.
+			// Check if the 'go-flutter' needs updates.
+			versioncheck.CheckFoGoFlutterUpdate(filepath.Join(wd, buildPath), currentTag)
+		}
+
+	} else {
+		fmt.Printf("hover: Downloading 'go-flutter' %s\n", buildBranch)
+
+		// when the buildBranch is set, fetch the go-flutter branch version.
+		err = upgradeGoFlutter(targetOS, engineCachePath)
+		if err != nil {
+			os.Exit(1)
+		}
 	}
 
 	var ldflags []string
@@ -237,7 +265,7 @@ func build(projectName string, targetOS string, vmArguments []string) {
 		fmt.Sprintf("-ldflags=%s", strings.Join(ldflags, " ")),
 		dotSlash+"cmd",
 	)
-	cmdGoBuild.Dir = filepath.Join(wd, "desktop")
+	cmdGoBuild.Dir = filepath.Join(wd, buildPath)
 	cmdGoBuild.Env = append(os.Environ(),
 		"GO111MODULE=on",
 		"CGO_LDFLAGS="+cgoLdflags,
@@ -246,6 +274,7 @@ func build(projectName string, targetOS string, vmArguments []string) {
 	cmdGoBuild.Stderr = os.Stderr
 	cmdGoBuild.Stdout = os.Stdout
 
+	fmt.Printf("hover: Compiling 'go-flutter' and plugins\n")
 	err = cmdGoBuild.Run()
 	if err != nil {
 		fmt.Printf("hover: Go build failed: %v\n", err)
